+++ storage/http_client.go
@@ -55,7 +55,7 @@ type httpStorageClient struct {
 	scheme                     string
 	settings                   *settings
 	config                     *storageConfig
-	dynamicReadReqStallTimeout *bucketDelay
+	dynamicReadReqStallTimeout *bucketDelayManager
 }
 
 // newHTTPStorageClient initializes a new storageClient that uses the HTTP-JSON
@@ -130,10 +130,10 @@ func newHTTPStorageClient(ctx context.Context, opts ...storageOption) (storageCl
 		return nil, fmt.Errorf("supplied endpoint %q is not valid: %w", ep, err)
 	}
 
-	var bd *bucketDelay
+	var bd *bucketDelayManager
 	if config.readStallTimeoutConfig != nil {
 		drrstConfig := config.readStallTimeoutConfig
-		bd, err = newBucketDelay(
+		bd, err = newBucketDelayManager(
 			drrstConfig.TargetPercentile,
 			getDynamicReadReqIncreaseRateFromEnv(),
 			getDynamicReadReqInitialTimeoutSecFromEnv(drrstConfig.Min),
@@ -1352,7 +1352,7 @@ func setRangeReaderHeaders(h http.Header, params *newRangeReaderParams) error {
 // readerReopen initiates a Read with offset and length, assuming we
 // have already read seen bytes.
 func readerReopen(ctx context.Context, header http.Header, params *newRangeReaderParams, s *settings,
-	doDownload func(context.Context) (*http.Response, error), applyConditions func() error, setGeneration func()) func(int64) (*http.Response, error) {
+		doDownload func(context.Context) (*http.Response, error), applyConditions func() error, setGeneration func()) func(int64) (*http.Response, error) {
 	return func(seen int64) (*http.Response, error) {
 		// If the context has already expired, return immediately without making a
 		// call.
@@ -1407,9 +1407,9 @@ func readerReopen(ctx context.Context, header http.Header, params *newRangeReade
 			}
 
 			partialContentNotSatisfied :=
-				!decompressiveTranscoding(res) &&
-					start > 0 && params.length != 0 &&
-					res.StatusCode != http.StatusPartialContent
+					!decompressiveTranscoding(res) &&
+							start > 0 && params.length != 0 &&
+							res.StatusCode != http.StatusPartialContent
 
 			if partialContentNotSatisfied {
 				res.Body.Close()
